Fork for dss-osc usage
This fork implements major changes to how the library executes tmux commands.
The original version requires that the output of your program is piped back in
to your shell, whereas this one uses os/exec to simply run the commands from the
context of the program.

This fork also currently lacks good tests, and, as such, should not be considered
production-ready. As exec.Cmd objects do not have an interface that can be
mocked to intercept the system calls, implementing proper tests will likely
require a major rewrite, to be completed at some point in the future.

Use at your own risk.